api/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==24.2.0
babel==2.16.0
bcrypt==5.0.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
defusedxml==0.7.1
discord.py==2.4.0
Django==5.1.5
djangorestframework==3.15.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
Flask==3.0.3
flask-babel==4.0.0
frozenlist==1.4.1
greenlet==3.2.3
h11==0.14.0
httpcore==1.0.7
httpie==3.2.4
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
motor==3.7.1
multidict==6.1.0
mysql-connector-python==9.0.0
mysqlclient==2.2.7
numpy==2.3.3
pandas==2.3.3
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.15.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9.1
setuptools==75.3.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.46.1
typer==0.15.2
typing_extensions==4.13.0
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
Werkzeug==3.0.4
yarl==1.11.1
